requirements.txt
fastapi
requests
axios
uvicorn
python-dotenv
firebase-admin